+++ igoMoney/igoMoney/Source/App/igoMoneyApp.swift
@@ -27,70 +27,3 @@ struct igoMoneyApp: App {
     }
   }
 }
-
-struct AppCore: Reducer {
-  enum State {
-    case logIn(MainCore.State)
-    case logOut(AuthCore.State)
-  }
-  
-  enum Action {
-    case mainAction(MainCore.Action)
-    case authAction(AuthCore.Action)
-  }
-  
-  var body: some Reducer<State, Action> {
-    Scope(state: /State.logIn, action: /Action.mainAction) {
-      MainCore()
-    }
-    
-    Scope(state: /State.logOut, action: /Action.authAction) {
-      AuthCore()
-    }
-    
-    Reduce { state, action in
-      switch action {
-      case .authAction(._presentMainScene):
-        state = .logIn(MainCore.State())
-        return .none
-      default:
-        return .none
-      }
-    }
-  }
-}
-
-struct AppView: View {
-  let store: StoreOf<AppCore>
-  
-  var body: some View {
-    SwitchStore(store) { state in
-      switch state {
-      case .logIn:
-        CaseLet(
-          /AppCore.State.logIn,
-          action: AppCore.Action.mainAction
-        ) { store in
-          MainScene(store: store)
-        }
-        
-      case .logOut:
-        CaseLet(
-          /AppCore.State.logOut,
-          action: AppCore.Action.authAction
-        ) { store in
-          AuthScene(store: store)
-        }
-      }
-    }
-  }
-}
-
-#Preview {
-  AppView(
-    store: Store(
-      initialState: AppCore.State.logIn(.init()),
-      reducer: { AppCore() }
-    )
-  )
-}
